+++ src/extensions/surveys/surveys-utils.tsx
@@ -582,7 +582,7 @@ export const shuffle = (array: any[]) => {
         .map((a) => a.value)
 }
 
-const reverseIfUnshuffled = (unshuffled: string[], shuffled: string[]): string[] => {
+const reverseIfUnshuffled = (unshuffled: any[], shuffled: any[]): any[] => {
     if (unshuffled.length === shuffled.length && unshuffled.every((val, index) => val === shuffled[index])) {
         return shuffled.reverse()
     }
@@ -622,7 +622,7 @@ export const getDisplayOrderQuestions = (survey: Survey): SurveyQuestion[] => {
         return survey.questions
     }
 
-    return shuffle(survey.questions)
+    return reverseIfUnshuffled(survey.questions, shuffle(survey.questions))
 }
 
 export const getSurveySeenKey = (survey: Survey): string => {