+++ src/posthog-surveys.ts
@@ -1,11 +1,13 @@
 import { PostHog } from './posthog-core'
 import { SURVEYS } from './constants'
-import { SurveyCallback, SurveyUrlMatchType } from './posthog-surveys-types'
+import { Survey, SurveyCallback, SurveyUrlMatchType } from './posthog-surveys-types'
 import { isUrlMatchingRegex } from './utils/request-utils'
-import { window, document, assignableWindow } from './utils/globals'
+import { SurveyEventReceiver } from './utils/survey-event-receiver'
+import { assignableWindow, document, window } from './utils/globals'
 import { DecideResponse } from './types'
 import { loadScript } from './utils'
 import { logger } from './utils/logger'
+import { isUndefined } from './utils/type-utils'
 
 export const surveyUrlValidationMap: Record<SurveyUrlMatchType, (conditionsUrl: string) => boolean> = {
     icontains: (conditionsUrl) =>
@@ -17,9 +19,13 @@ export const surveyUrlValidationMap: Record<SurveyUrlMatchType, (conditionsUrl: 
 export class PostHogSurveys {
     instance: PostHog
     private _decideServerResponse?: boolean
+    public _surveyEventReceiver: SurveyEventReceiver | null
 
     constructor(instance: PostHog) {
         this.instance = instance
+        // we set this to undefined here because we need the persistence storage for this type
+        // but that's not initialized until loadIfEnabled is called.
+        this._surveyEventReceiver = null
     }
 
     afterDecideResponse(response: DecideResponse) {
@@ -31,6 +37,9 @@ export class PostHogSurveys {
         const surveysGenerator = assignableWindow?.extendPostHogWithSurveys
 
         if (!this.instance.config.disable_surveys && this._decideServerResponse && !surveysGenerator) {
+            if (this._surveyEventReceiver == null) {
+                this._surveyEventReceiver = new SurveyEventReceiver(this.instance.persistence)
+            }
             loadScript(this.instance.requestRouter.endpointFor('assets', '/static/surveys.js'), (err) => {
                 if (err) {
                     return logger.error(`Could not load surveys script`, err)
@@ -47,6 +56,11 @@ export class PostHogSurveys {
         if (this.instance.config.disable_surveys) {
             return callback([])
         }
+
+        if (this._surveyEventReceiver == null) {
+            this._surveyEventReceiver = new SurveyEventReceiver(this.instance.persistence)
+        }
+
         const existingSurveys = this.instance.get_property(SURVEYS)
         if (!existingSurveys || forceReload) {
             this.instance._send_request({
@@ -61,6 +75,22 @@ export class PostHogSurveys {
                         return callback([])
                     }
                     const surveys = response.json.surveys || []
+
+                    const eventBasedSurveys = surveys.filter(
+                        (survey: Survey) =>
+                            survey.conditions?.events &&
+                            survey.conditions?.events?.values &&
+                            survey.conditions?.events?.values?.length > 0
+                    )
+
+                    if (eventBasedSurveys.length > 0 && !isUndefined(this.instance._addCaptureHook)) {
+                        this._surveyEventReceiver?.register(eventBasedSurveys)
+                        const onEventName = (eventName: string) => {
+                            this._surveyEventReceiver?.on(eventName)
+                        }
+                        this.instance._addCaptureHook(onEventName)
+                    }
+
                     this.instance.persistence?.register({ [SURVEYS]: surveys })
                     return callback(surveys)
                 },
@@ -75,6 +105,7 @@ export class PostHogSurveys {
             const activeSurveys = surveys.filter((survey) => {
                 return !!(survey.start_date && !survey.end_date)
             })
+
             const conditionMatchedSurveys = activeSurveys.filter((survey) => {
                 if (!survey.conditions) {
                     return true
@@ -89,6 +120,9 @@ export class PostHogSurveys {
                     : true
                 return urlCheck && selectorCheck
             })
+
+            // get all the surveys that have been activated so far with user actions.
+            const activatedSurveys: string[] | undefined = this._surveyEventReceiver?.getSurveys()
             const targetingMatchedSurveys = conditionMatchedSurveys.filter((survey) => {
                 if (!survey.linked_flag_key && !survey.targeting_flag_key && !survey.internal_targeting_flag_key) {
                     return true
@@ -104,7 +138,14 @@ export class PostHogSurveys {
                     ? this.instance.featureFlags.isFeatureEnabled(survey.internal_targeting_flag_key)
                     : true
 
-                return linkedFlagCheck && targetingFlagCheck && internalTargetingFlagCheck
+                const hasEvents =
+                    survey.conditions?.events &&
+                    survey.conditions?.events?.values &&
+                    survey.conditions?.events?.values.length > 0
+                const eventBasedTargetingFlagCheck = hasEvents ? activatedSurveys?.includes(survey.id) : true
+                return (
+                    linkedFlagCheck && targetingFlagCheck && internalTargetingFlagCheck && eventBasedTargetingFlagCheck
+                )
             })
 
             return callback(targetingMatchedSurveys)